Refactor fetchUnapprovedUsers function with improved structure and comments
Related to #1517
